+++ components/notification-drawer/components/Drawer.tsx
@@ -10,20 +10,14 @@ const { FSBL } = window;
 function Drawer(props: Props): React.ReactElement {
 	const [animationClass, setAnimationClass] = useState("slide-in-right");
 
-	window.onblur = () => {
-		setAnimationClass("slide-out-right");
-		console.log("* window blurred");
-	};
-
 	window.onfocus = () => {
+		setAnimationClass("");
 		setAnimationClass("slide-in-right");
-		console.log("* window has focus");
 	};
-
 	const animationComplete = () => {
 		animationClass === "slide-out-right" &&
-			// eslint-disable-next-line @typescript-eslint/no-empty-function
-			FSBL.Clients.WindowClient.minimize(() => {});
+			// TODO: remove this and change for hide
+			FSBL.Clients.WindowClient.minimize();
 	};
 	return (
 		<div id="drawer" className={animationClass} onAnimationEnd={animationComplete}>
@@ -31,6 +25,7 @@ function Drawer(props: Props): React.ReactElement {
 				<img
 					src="../shared/assets/dashboard.svg"
 					id="notification-center-icon"
+					// TODO: Move this out
 					onClick={() =>
 						FSBL.Clients.LauncherClient.showWindow(
 							{