+++ components/notification-toasts/components/Drawer.tsx
@@ -1,9 +1,10 @@
 import * as React from "react";
+const _get = require("lodash.get");
 import useNotifications from "../../shared/hooks/useNotifications";
 import INotification from "../../../types/Notification-definitions/INotification";
 import { SpawnParams } from "../../../types/FSBL-definitions/services/window/Launcher/launcher";
 
-const { useEffect, useRef } = React;
+const { useEffect, useRef, useState } = React;
 
 interface Props {
 	children: React.PropsWithChildren<any>;
@@ -12,14 +13,39 @@ interface Props {
 }
 
 function Drawer(props: Props): React.ReactElement {
-	const { setNotificationDrawerPosition, minimizeWindow } = useNotifications();
+	const { setNotificationDrawerPosition, minimizeWindow: hideWindow, activeNotifications } = useNotifications();
 	const inputEl = useRef(null);
 	const { notifications, windowShowParams } = props;
-
+	const [monitor, setMonitor] = useState(null);
 	useEffect(() => {
-		windowShowParams.height = inputEl.current.getBoundingClientRect().height;
-		notifications.length === 0 ? minimizeWindow() : setNotificationDrawerPosition(windowShowParams);
-	}, [minimizeWindow, notifications, setNotificationDrawerPosition, windowShowParams]);
+		const resizeDrawerHeight = async () => {
+			if (!monitor) {
+				const monitorInfo = await FSBL.Clients.LauncherClient.getMonitorInfo({
+					monitor: "primary"
+				});
+				setMonitor(monitorInfo);
+			}
+			const { height: monitorHeight } = _get(monitor, "data.availableRect", 3000);
+			const notificationHeight = 162;
+
+			const { width } = FSBL.Clients.WindowClient.options.customData.window;
+
+			windowShowParams.height = notificationHeight * activeNotifications(notifications).length;
+			windowShowParams.width = width;
+
+			if (windowShowParams.height === 0) {
+				windowShowParams.height = 1;
+				windowShowParams.width = 1;
+			}
+
+			if (windowShowParams.height > monitorHeight) {
+				windowShowParams.height = monitorHeight;
+			}
+
+			await setNotificationDrawerPosition(windowShowParams); // sets the window size / height
+		};
+		resizeDrawerHeight();
+	}, [monitor, notifications, setNotificationDrawerPosition, windowShowParams]);
 
 	return (
 		<div id="toasts-drawer" ref={inputEl}>