+++ components/shared/components/icons/NotificationIcon.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
-import SVGBase from "./SVGBase";
+import SVGBase, { SVGProps } from "./SVGBase";
 
-export default function NotificationIcon(props) {
+export default function NotificationIcon(props: SVGProps) {
 	return (
 		<SVGBase
 			{...props}